requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12 
cycler==0.11.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
flatbuffers==2.0
fonttools==4.36.0
gast==0.5.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.3
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.20.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image==0.17.2
opencv-python
tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
wldhx.yadisk-direct